requirements.txt
pymongo==4.*
requests==2.*
asyncio==3.*
pytz==2022.*
discord.py==1.*